입점 신청 기능 수정 및 추가

+++ idwith/src/main/java/com/idwith/mpweb/user/service/SellerCheckService.java
@@ -1,5 +1,5 @@
 package com.idwith.mpweb.user.service;
 
 public interface SellerCheckService {
-	int getSeller(String user_id);
+	Long getSeller(String user_id);
 }
